Add details for "About these changes" for Volunteer Only
Will show up in email like:
```
You submitted feedback suggesting that the entry
The position is awesome (Pwi Di Goo Zing Ne Yaa Zhing Advisory Services)
be updated in Volunteer Database More Data

The following is a summary of the suggested changes to field(s) in the database:

About These Changes: This is not a complete update. I have only reviewed or modified some of the information in this record.

Please note that the suggested changes will not be added until they are reviewed by our data management staff.

You can view the record at:
http://fizban-current/volunteer/record/V-GOO0043?Ln=en-CA
```