Fix VOL Profile links show when not enabled in view
The old format volunteer search form had a guard to prevent the
volunteer profiles links from being added if not configured in the view
but there was no construct to allow for that in more recent templates.

This change introduces the new `USE_VOL_PROFILES_VIEW` template variable
which encapsulates the logic from the old form to be used in newer
templates and applies that to a new if condition on the VOL Search Dual
Panel and Multi Panel templates.

See [here](
https://github.com/OpenCIOC/onlineresources/blob/master/python/cioc/web/vol/templates/bsearch.mak#L40
) for the old logic.