Fix date abbreviations don't match SQL Server/Windows ones
Upgrading to babel 2.10.3 brought CLDR to version 41 and the
abbrviations don't match the other sources of date strings if they are a
CA variant. This uses the base abbreviations for en-CA and fr-CA (i.e.
en and fr) instead which is consistent with the dates formatted by SQL
Server and ASP/vbscript.